+++ src/main/java/nl/nn/testtool/storage/proofofmigration/ProofOfMigrationStorage.java
@@ -19,10 +19,14 @@ import java.lang.invoke.MethodHandles;
 import java.sql.Types;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
+
+import javax.inject.Inject;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
 
 import lombok.Setter;
 import nl.nn.testtool.Checkpoint;
@@ -42,9 +46,74 @@ import nl.nn.testtool.storage.database.DatabaseStorage;
  * 
  * @author Jaco de Groot
  */
+//@Dependent disabled for Quarkus for now because of the use of JdbcTemplate
 public class ProofOfMigrationStorage extends DatabaseStorage {
 	private static Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-	private @Setter TestTool testTool;
+	private static final String GROUP_BY = "group by COMPONENT, CORRELATION_ID";
+	private @Setter @Inject @Autowired TestTool testTool;
+	private @Setter List<String> columns;
+	private @Setter boolean showErrorsOnly = false;
+
+	@Override
+	public String getName() {
+		if (name == null) {
+			return "proofOfMigrationStorage";
+		} else {
+			return name;
+		}
+	}
+
+	public String getTable() {
+		if (table == null) {
+			return "PROOF_OF_MIGRATION";
+		} else {
+			return table;
+		}
+	}
+
+	@Override
+	public List<String> getMetadataNames() {
+		return getColumns();
+	}
+
+	public List<String> getColumns() {
+		if (columns == null) {
+			return new ArrayList<String>(Arrays.asList(
+					"ID", "TIMESTAMP", "COMPONENT", "ENDPOINT_NAME", "CORRELATION_ID", "NR OF CHECKPOINTS", "STATUS"));
+		} else {
+			return columns;
+		}
+	}
+
+	public String getStorageIdColumn() {
+		if (storageIdColumn == null) {
+			return "ID";
+		} else {
+			return storageIdColumn;
+		}
+	}
+
+	@Override
+	public List<String> getIntegerColumns() {
+		if (integerColumns == null) {
+			return new ArrayList<String>(Arrays.asList("ID", "CHECKPOINT_NR"));
+		} else {
+			return integerColumns;
+		}
+	}
+
+	@Override
+	public List<String> getTimestampColumns() {
+		if (timestampColumns == null) {
+			return new ArrayList<String>(Arrays.asList("TIMESTAMP"));
+		} else {
+			return timestampColumns;
+		}
+	}
+
+	public boolean isShowErrorsOnly() {
+		return showErrorsOnly;
+	}
 
 	@Override
 	protected void buildMetadataQuery(int maxNumberOfRecords, List<String> metadataNames, List<String> searchValues,
@@ -54,15 +123,67 @@ public class ProofOfMigrationStorage extends DatabaseStorage {
 				query, args, argTypes);
 		if (log.isDebugEnabled()) {
 			log.debug("Get metadata original query: " + query.toString());
-			// E.g.: select ID, TIMESTAMP, COMPONENT, CORRELATION_ID, NR OF CHECKPOINTS, STATUS from PROOF_OF_MIGRATION order by ID desc limit 10
+			// E.g.: select ID, TIMESTAMP, COMPONENT, ENDPOINT_NAME, CORRELATION_ID, NR OF CHECKPOINTS, STATUS from PROOF_OF_MIGRATION order by ID desc limit 10
 			// The new metadata query will be logged by the super class
-			// E.g.: select min(ID) as min_id, min(TIMESTAMP), COMPONENT, CORRELATION_ID, max(CHECKPOINT_NR), min(STATUS) from PROOF_OF_MIGRATION group by COMPONENT, CORRELATION_ID order by min_id desc limit 10
+			// E.g.: select min(ID) as min_id, min(TIMESTAMP), COMPONENT, min(concat(CHECKPOINT_NR, '. ', ENDPOINT_NAME)), CORRELATION_ID, count(CHECKPOINT_NR), min(STATUS) from PROOF_OF_MIGRATION where STATUS != 'Success' group by COMPONENT, CORRELATION_ID order by min_id desc limit 10
 		}
 		replace(query, "ID", "min(ID) as min_id");
 		replace(query, "TIMESTAMP", "min(TIMESTAMP)");
-		replace(query, "STATUS", "min(STATUS)");
+		replace(query, "ENDPOINT_NAME", "min(concat(CHECKPOINT_NR, '. ', ENDPOINT_NAME))");
 		replace(query, "NR OF CHECKPOINTS", "count(CHECKPOINT_NR)");
-		replace(query, "order by ID" , "group by COMPONENT, CORRELATION_ID order by min_id");
+		replace(query, "STATUS", "min(STATUS)");
+		replace(query, "order by ID" , GROUP_BY + " order by min_id");
+		if (isShowErrorsOnly()) {
+			replace(query, "group by" , "where STATUS != 'Success' group by");
+		}
+	}
+
+	/*
+	 * Make sure the record for old component and the record for the new component (that have the same correlation id)
+	 * are displayed directly below each other. The order by ID in the query doesn't guarantee that.
+	 */
+	@Override
+	protected void postProcessMetadataResult(List<List<Object>> metadata, int maxNumberOfRecords,
+			List<String> metadataNames, List<String> searchValues, int metadataValueType) {
+		int correlationIdPosition = -1;
+		int position = 0;
+		for (String metadataName : metadataNames) {
+			if (metadataName.equals("CORRELATION_ID")) {
+				correlationIdPosition = position;
+			}
+			position++;
+		}
+		if (correlationIdPosition == -1) {
+			log.warn("Could not find field CORRELATION_ID");
+		} else {
+			for (int i = metadata.size() - 1; i > 0; i--) {
+				Object correlationId = metadata.get(i).get(correlationIdPosition);
+				if (!correlationId.equals(metadata.get(i - 1).get(correlationIdPosition))
+						&& !(i < metadata.size() - 1
+								&& correlationId.equals(metadata.get(i + 1).get(correlationIdPosition)))) {
+					for (int j = i - 2; j > -1; j--) {
+						if (correlationId.equals(metadata.get(j).get(correlationIdPosition))) {
+							List<Object> record = metadata.remove(j);
+							metadata.add(i - 1, record);
+							j = 0;
+						}
+					}
+				}
+			}
+		}
+	}
+
+	@Override
+	protected void buildSizeQuery(StringBuilder query) throws StorageException {
+		super.buildSizeQuery(query);
+		if (log.isDebugEnabled()) {
+			log.debug("Get size original query: " + query.toString());
+			// E.g.: select count(*) from PROOF_OF_MIGRATION
+			// The new size query will be logged by the super class
+			// E.g.: select count(*) from (select COMPONENT, CORRELATION_ID from PROOF_OF_MIGRATION group by COMPONENT, CORRELATION_ID) as DUMMY
+		}
+		replace(query, "from", "from (select " + GROUP_BY.substring(9) + " from");
+		query.append(" " + GROUP_BY + ") as DUMMY");
 	}
 
 	@Override
@@ -70,8 +191,8 @@ public class ProofOfMigrationStorage extends DatabaseStorage {
 		String query =
 				"select CORRELATION_ID, COMPONENT, CHECKPOINT_NR, ENDPOINT_NAME, ACTION, MESSAGE, TIMESTAMP from "
 				+ getTable() + " where CORRELATION_ID = (select CORRELATION_ID from " + getTable() + " where "
-				+ getStorageIdColumnName() + " = ?) and COMPONENT = (select COMPONENT from " + getTable() + " where "
-				+ getStorageIdColumnName() + " = ?) order by CHECKPOINT_NR";
+				+ getStorageIdColumn() + " = ?) and COMPONENT = (select COMPONENT from " + getTable() + " where "
+				+ getStorageIdColumn() + " = ?) order by CHECKPOINT_NR";
 		log.debug("Get report query: " + query);
 		List<List<Object>> result = getJdbcTemplate().query(query, new Object[]{storageId, storageId},
 				new int[] {Types.INTEGER, Types.INTEGER},